Switch StytchProductConfig items to use Int instead of UInt because Moshi doesn't like them; Update return type for Event logging since it returns a 204; Update apis to send lists of events, since that's what web-backend expects

+++ ui/src/main/kotlin/com/stytch/sdk/ui/screens/NewUserScreen.kt
@@ -64,7 +64,7 @@ internal object NewUserScreen : AndroidScreen(), Parcelable {
             onEmailAddressChanged = viewModel::onEmailAddressChanged,
             onPasswordChanged = viewModel::onPasswordChanged,
             onEmailAndPasswordSubmitted = {
-                viewModel.createAccountWithPassword(productConfig.sessionOptions.sessionDurationMinutes)
+                viewModel.createAccountWithPassword(productConfig.sessionOptions.sessionDurationMinutes.toUInt())
             },
             hasEML = hasEML,
             hasEmailOTP = hasEmailOTP,
